web/backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
opencv-python
meshlib
Pillow
numpy
python-multipart
